requirements.txt
async-timeout==5.0.1
certifi==2026.1.4
charset-normalizer==3.4.4
idna==3.11
python-dotenv==1.2.1
redis==7.0.1
requests==2.32.5
spotipy==2.25.2
urllib3==2.6.3
ytmusicapi==1.10.3
